% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheit converter}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{Numeric input indicating temperature in Fahrenheit}
}
\value{
Numeric representing \code{temp_F} converted to Celsius.
}
\description{
This function converts temp from Fahrenheit to Celsius
}
\examples{
f_to_c(12)
f_to_c(200)

}
\keyword{temperature}
